remove the skl prefix to sklearn converters remove the onnx prefix to onnx converters move lightgbm and xgb converters out of the sklearn folder
