+++ doc/html/hummingbird/ml/operator_converters/onnx/index.html
@@ -43,31 +43,31 @@ All onnx operators converters are stored under this package.
 <section>
 <h2 class="section-title" id="header-submodules">Sub-modules</h2>
 <dl>
+<dt><code class="name"><a title="hummingbird.ml.operator_converters.onnx.array_feature_extractor" href="array_feature_extractor.html">hummingbird.ml.operator_converters.onnx.array_feature_extractor</a></code></dt>
+<dd>
+<div class="desc"><p>Converter for ONNX-ML Array Feature Extractor.</p></div>
+</dd>
+<dt><code class="name"><a title="hummingbird.ml.operator_converters.onnx.linear" href="linear.html">hummingbird.ml.operator_converters.onnx.linear</a></code></dt>
+<dd>
+<div class="desc"><p>Converters for ONNX-ML linear models.</p></div>
+</dd>
+<dt><code class="name"><a title="hummingbird.ml.operator_converters.onnx.normalizer" href="normalizer.html">hummingbird.ml.operator_converters.onnx.normalizer</a></code></dt>
+<dd>
+<div class="desc"></div>
+</dd>
+<dt><code class="name"><a title="hummingbird.ml.operator_converters.onnx.one_hot_encoder" href="one_hot_encoder.html">hummingbird.ml.operator_converters.onnx.one_hot_encoder</a></code></dt>
+<dd>
+<div class="desc"><p>Converter for ONNX-ML One Hot Encoder.</p></div>
+</dd>
 <dt><code class="name"><a title="hummingbird.ml.operator_converters.onnx.onnx_operator" href="onnx_operator.html">hummingbird.ml.operator_converters.onnx.onnx_operator</a></code></dt>
 <dd>
 <div class="desc"><p>Converters for ONNX operators.</p></div>
 </dd>
-<dt><code class="name"><a title="hummingbird.ml.operator_converters.onnx.onnxml_array_feature_extractor" href="onnxml_array_feature_extractor.html">hummingbird.ml.operator_converters.onnx.onnxml_array_feature_extractor</a></code></dt>
-<dd>
-<div class="desc"><p>Converter for ONNX-ML Array Feature Extractor.</p></div>
-</dd>
-<dt><code class="name"><a title="hummingbird.ml.operator_converters.onnx.onnxml_linear" href="onnxml_linear.html">hummingbird.ml.operator_converters.onnx.onnxml_linear</a></code></dt>
-<dd>
-<div class="desc"><p>Converters for ONNX-ML linear models.</p></div>
-</dd>
-<dt><code class="name"><a title="hummingbird.ml.operator_converters.onnx.onnxml_normalizer" href="onnxml_normalizer.html">hummingbird.ml.operator_converters.onnx.onnxml_normalizer</a></code></dt>
-<dd>
-<div class="desc"></div>
-</dd>
-<dt><code class="name"><a title="hummingbird.ml.operator_converters.onnx.onnxml_one_hot_encoder" href="onnxml_one_hot_encoder.html">hummingbird.ml.operator_converters.onnx.onnxml_one_hot_encoder</a></code></dt>
-<dd>
-<div class="desc"><p>Converter for ONNX-ML One Hot Encoder.</p></div>
-</dd>
-<dt><code class="name"><a title="hummingbird.ml.operator_converters.onnx.onnxml_scaler" href="onnxml_scaler.html">hummingbird.ml.operator_converters.onnx.onnxml_scaler</a></code></dt>
+<dt><code class="name"><a title="hummingbird.ml.operator_converters.onnx.scaler" href="scaler.html">hummingbird.ml.operator_converters.onnx.scaler</a></code></dt>
 <dd>
 <div class="desc"><p>Converter for ONNX-ML Scaler.</p></div>
 </dd>
-<dt><code class="name"><a title="hummingbird.ml.operator_converters.onnx.onnxml_tree_ensemble" href="onnxml_tree_ensemble.html">hummingbird.ml.operator_converters.onnx.onnxml_tree_ensemble</a></code></dt>
+<dt><code class="name"><a title="hummingbird.ml.operator_converters.onnx.tree_ensemble" href="tree_ensemble.html">hummingbird.ml.operator_converters.onnx.tree_ensemble</a></code></dt>
 <dd>
 <div class="desc"><p>Converters for ONNX-ML tree-ensemble models.</p></div>
 </dd>
@@ -97,13 +97,13 @@ All onnx operators converters are stored under this package.
 </li>
 <li><h3><a href="#header-submodules">Sub-modules</a></h3>
 <ul>
+<li><code><a title="hummingbird.ml.operator_converters.onnx.array_feature_extractor" href="array_feature_extractor.html">hummingbird.ml.operator_converters.onnx.array_feature_extractor</a></code></li>
+<li><code><a title="hummingbird.ml.operator_converters.onnx.linear" href="linear.html">hummingbird.ml.operator_converters.onnx.linear</a></code></li>
+<li><code><a title="hummingbird.ml.operator_converters.onnx.normalizer" href="normalizer.html">hummingbird.ml.operator_converters.onnx.normalizer</a></code></li>
+<li><code><a title="hummingbird.ml.operator_converters.onnx.one_hot_encoder" href="one_hot_encoder.html">hummingbird.ml.operator_converters.onnx.one_hot_encoder</a></code></li>
 <li><code><a title="hummingbird.ml.operator_converters.onnx.onnx_operator" href="onnx_operator.html">hummingbird.ml.operator_converters.onnx.onnx_operator</a></code></li>
-<li><code><a title="hummingbird.ml.operator_converters.onnx.onnxml_array_feature_extractor" href="onnxml_array_feature_extractor.html">hummingbird.ml.operator_converters.onnx.onnxml_array_feature_extractor</a></code></li>
-<li><code><a title="hummingbird.ml.operator_converters.onnx.onnxml_linear" href="onnxml_linear.html">hummingbird.ml.operator_converters.onnx.onnxml_linear</a></code></li>
-<li><code><a title="hummingbird.ml.operator_converters.onnx.onnxml_normalizer" href="onnxml_normalizer.html">hummingbird.ml.operator_converters.onnx.onnxml_normalizer</a></code></li>
-<li><code><a title="hummingbird.ml.operator_converters.onnx.onnxml_one_hot_encoder" href="onnxml_one_hot_encoder.html">hummingbird.ml.operator_converters.onnx.onnxml_one_hot_encoder</a></code></li>
-<li><code><a title="hummingbird.ml.operator_converters.onnx.onnxml_scaler" href="onnxml_scaler.html">hummingbird.ml.operator_converters.onnx.onnxml_scaler</a></code></li>
-<li><code><a title="hummingbird.ml.operator_converters.onnx.onnxml_tree_ensemble" href="onnxml_tree_ensemble.html">hummingbird.ml.operator_converters.onnx.onnxml_tree_ensemble</a></code></li>
+<li><code><a title="hummingbird.ml.operator_converters.onnx.scaler" href="scaler.html">hummingbird.ml.operator_converters.onnx.scaler</a></code></li>
+<li><code><a title="hummingbird.ml.operator_converters.onnx.tree_ensemble" href="tree_ensemble.html">hummingbird.ml.operator_converters.onnx.tree_ensemble</a></code></li>
 </ul>
 </li>
 </ul>
