+++ hummingbird/ml/operator_converters/__init__.py
@@ -18,23 +18,23 @@ constants = _Constants(converter_constants, hummingbird_constants)
 
 # To register a converter for scikit-learn API operators, import associated modules here.
 from .onnx import onnx_operator  # noqa: E402
-from .onnx import onnxml_array_feature_extractor  # noqa: E402
-from .onnx import onnxml_linear  # noqa: E402
-from .onnx import onnxml_normalizer  # noqa: E402
-from .onnx import onnxml_one_hot_encoder  # noqa: E402
-from .onnx import onnxml_scaler  # noqa: E402
-from .onnx import onnxml_tree_ensemble  # noqa: E402
-from .sklearn import lightgbm  # noqa: E402
-from .sklearn import skl_array_feature_extractor  # noqa: E402
-from .sklearn import skl_decision_tree  # noqa: E402
-from .sklearn import skl_gbdt  # noqa: E402
-from .sklearn import skl_iforest  # noqa: E402
-from .sklearn import skl_linear  # noqa: E402
-from .sklearn import skl_normalizer  # noqa: E402
-from .sklearn import skl_one_hot_encoder  # noqa: E402
-from .sklearn import skl_scaler  # noqa: E402
-from .sklearn import skl_sv  # noqa: E402
-from .sklearn import xgb  # noqa: E402
+from .onnx import array_feature_extractor as onnx_afe  # noqa: E402, F811
+from .onnx import linear as onnx_linear  # noqa: E402, F811
+from .onnx import normalizer as onnx_normalizer  # noqa: E402, F811
+from .onnx import one_hot_encoder as onnx_ohe  # noqa: E402, F811
+from .onnx import scaler as onnx_scaler  # noqa: E402, F811
+from .onnx import tree_ensemble  # noqa: E402
+from .sklearn import array_feature_extractor as sklearn_afe  # noqa: E402
+from .sklearn import decision_tree  # noqa: E402
+from .sklearn import gbdt  # noqa: E402
+from .sklearn import iforest  # noqa: E402
+from .sklearn import linear as sklearn_linear  # noqa: E402
+from .sklearn import normalizer as sklearn_normalizer  # noqa: E402
+from .sklearn import one_hot_encoder as sklearn_ohe  # noqa: E402
+from .sklearn import scaler as sklearn_scaler  # noqa: E402
+from .sklearn import sv  # noqa: E402
+from . import lightgbm  # noqa: E402
+from . import xgb  # noqa: E402
 
 
 __pdoc__ = {}
