+++ hummingbird/ml/operator_converters/sklearn/array_feature_extractor.py
